Move device_class
