+++ custom_components/tahoma/cover_devices/vertical_cover.py
@@ -1,5 +1,14 @@
 """Support for TaHoma cover - shutters etc."""
-from homeassistant.components.cover import ATTR_POSITION
+from homeassistant.components.cover import (
+    ATTR_POSITION,
+    DEVICE_CLASS_AWNING,
+    DEVICE_CLASS_BLIND,
+    DEVICE_CLASS_CURTAIN,
+    DEVICE_CLASS_GARAGE,
+    DEVICE_CLASS_GATE,
+    DEVICE_CLASS_SHUTTER,
+    DEVICE_CLASS_WINDOW,
+)
 
 from custom_components.tahoma.cover_devices.tahoma_cover import TahomaCover
 
@@ -24,9 +33,33 @@ SERVICE_COVER_POSITION_LOW_SPEED = "set_cover_position_low_speed"
 
 SUPPORT_COVER_POSITION_LOW_SPEED = 1024
 
+TAHOMA_COVER_DEVICE_CLASSES = {
+    "Blind": DEVICE_CLASS_BLIND,
+    "Curtain": DEVICE_CLASS_CURTAIN,
+    "ExteriorScreen": DEVICE_CLASS_BLIND,
+    "ExteriorVenetianBlind": DEVICE_CLASS_BLIND,
+    "GarageDoor": DEVICE_CLASS_GARAGE,
+    "Gate": DEVICE_CLASS_GATE,
+    "MyFoxSecurityCamera": DEVICE_CLASS_SHUTTER,
+    "Pergola": DEVICE_CLASS_AWNING,
+    "RollerShutter": DEVICE_CLASS_SHUTTER,
+    "SwingingShutter": DEVICE_CLASS_SHUTTER,
+    "VeluxInteriorBlind": DEVICE_CLASS_BLIND,
+    "Window": DEVICE_CLASS_WINDOW,
+}
+
 
 class TahomaVerticalCover(TahomaCover):
     """Representation of a TaHoma Cover."""
+
+    @property
+    def device_class(self):
+        """Return the class of the device."""
+        return (
+            TAHOMA_COVER_DEVICE_CLASSES.get(self.device.widget)
+            or TAHOMA_COVER_DEVICE_CLASSES.get(self.device.ui_class)
+            or DEVICE_CLASS_BLIND
+        )
 
     @property
     def current_cover_position(self):
